requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
distro==1.8.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
openai==1.3.5
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-frontmatter==1.0.1
PyYAML==6.0.1
sniffio==1.3.0
sqlite-vss==0.1.2
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.8.0
